feat: feat/add-trivy
