+++ go/security/tests.go
@@ -6,16 +6,6 @@ package security
 
 import "strings"
 
-// TrivyReport represents the structure of a Trivy JSON report
-type TrivyReport struct {
-	Results []struct {
-		Vulnerabilities []struct {
-			VulnerabilityID string `json:"VulnerabilityID"`
-			Severity        string `json:"Severity"`
-		} `json:"Vulnerabilities"`
-	} `json:"Results"`
-}
-
 // Helper function to extract coverage from test output
 func ExtractCoverage(testOutput string) string {
 	// Look for a line like "coverage: 75.0% of statements"
