+++ trivy/dagger.json
@@ -1,0 +1,14 @@
+{
+  "name": "trivy",
+  "engineVersion": "v0.18.12",
+  "sdk": {
+    "source": "go"
+  },
+  "dependencies": [
+    {
+      "name": "trivy",
+      "source": "github.com/luanmtruong/daggerverse/trivy",
+      "pin": "96c13b929c636316317f745ff36cda4e4c66f680"
+    }
+  ]
+}
